Revert "product optimize route added"
This reverts commit 0a223c2a2cb171e2e119e031dfb8a725adf8993e.

# Conflicts:
#	v2Server/Notebooks/vectara indexing_querying.ipynb
