+++ v2Server/.env_example
@@ -1,9 +1,4 @@
 
 <!-- Rename the .env_example file to .env and replace you API keys here -->
 
-VECTARA_CUSTOMER_ID=""
-VECTARA_CLIENT_ID=""
-VECTARA_CLIENT_SECRET=""
-VECTARA_CORPUS_ID=""
-VECTARA_API_KEY=""
-OPENAI_APi_KEY=""
+V_API_KEY = YOUR_VECTARA_API_KEY
