+++ v2Server/server.py
@@ -1,0 +1,11 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+@app.route('/')
+def wellcome():
+    return "<center><h1> Wellcome to EcoFector<h1></center>"
+
+
+if __name__ == '__main__':
+    app.run(debug=True, host='0.0.0.0')
