+++ v2Server/vectara_RAG/.env_sample
@@ -1,0 +1,7 @@
+VECTARA_CUSTOMER_ID=""
+VECTARA_CLIENT_ID=""
+VECTARA_CLIENT_SECRET=""
+VECTARA_CORPUS_ID=""
+VECTARA_API_KEY=""
+OPENAI_APi_KEY=""
+
