+++ v2Server/vectara_RAG/requirements.txt
@@ -1,0 +1,3 @@
+openai==1.16.1
+python-dotenv==1.0.1
+vectara==0.0.1
